remove navbar and move all components to single page with simpler directory nesting
