+++ src/components/Contact.tsx
@@ -4,7 +4,7 @@ export default function Contact(): JSX.Element {
     return (
         <div className='contact'>
             <div className='body'>
-                <h1>Let's Chat!</h1>
+                <h2>Let's Chat!</h2>
                 <div className="logo-container" id="contact-logos">
                     <i className="fa-solid fa-mobile fa-2x"> 7 0 3. 6 2 5. 9 5 9 5</i><br/><br/>
                     <i className="fa-solid fa-envelope fa-2x"> b m i t c h u m. d e v @ g m a i l. c o m</i>
